dj_mls/requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
cryptography==1.9
Django==1.11.3
django-cors-headers==2.1.0
django-extensions==1.8.1
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.3
enum34==1.1.6
Fabric==1.13.2
idna==2.5
ipaddress==1.0.18
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
openapi-codec==1.3.1
paramiko==2.2.1
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.18
PyNaCl==1.1.2
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
